Attribution
Hi @liantze 
Thats a really cool project, I'm a dev at https://github.com/enhancv and we've noticed that you're using our design for it. We don't mind, really happy about it actually, but it would be great if you could put some kind of attribution somewhere in the README.md. I've made a pull request with a suggestion, but wherever you put it it'll be great. (Also we're planning on putting better copyright massages on those resumes in the future as its a bit confusing now)